move SerializedTransaction to the right module
